+++ app/(base-layout)/layout.tsx
@@ -1,5 +1,5 @@
 import { ReduxProvider } from '@/app/redux/redux-provider'
-import { withIPRestriction } from '@/shared/lib/withIPRestriction'
+import { withIPRestriction } from '@/shared/lib/with-iP-restriction'
 import { BaseLayout } from '@/shared/ui/base-layout'
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
